+++ packages/ui-causal-graph-editor/src/index.tsx
@@ -30,9 +30,6 @@ export {
     SpringLayout,
     FcoseLayout,
     ForceAtlasLayout,
-    TieredGraphLayoutBuilder,
-    TiersConfig,
-    GraphTiers,
 } from './shared/graph-layout';
 export { default as GraphContext } from './shared/graph-context';
 export { causalGraphParser } from './shared/parsers';
