+++ packages/ui-causal-graph-editor/src/shared/graph-layout/common.tsx
@@ -18,8 +18,6 @@ import { LayoutMapping, XYPosition } from 'graphology-layout/utils';
 
 import { SimulationGraph } from '../../types';
 
-export type DirectionType = 'horizontal' | 'vertical';
-
 export abstract class GraphLayoutBuilder<T> {
     _nodeSize = 64;
 
@@ -46,22 +44,6 @@ export abstract class GraphLayoutBuilder<T> {
     }
 
     abstract build(): T;
-}
-
-export interface TiersConfig {
-    group: string;
-    order_nodes_by?: string;
-    rank?: string[];
-}
-
-export type GraphTiers = string[][] | TiersConfig;
-
-/**
- * Defines necessary properties that need to be implemented by graph layouts to support tiered layouts
- */
-export interface TieredGraphLayoutBuilder {
-    orientation: DirectionType;
-    tiers: GraphTiers;
 }
 
 /**
