+++ packages/ui-causal-graph-editor/src/shared/graph-layout/planar-layout.tsx
@@ -17,9 +17,9 @@
 import { coordQuad, decrossTwoLayer, layeringSimplex, sugiyama } from 'd3-dag';
 import { LayoutMapping, XYPosition } from 'graphology-layout/utils';
 
-import { SimulationGraph } from '../../types';
+import { DirectionType, GraphTiers, SimulationGraph, TieredGraphLayoutBuilder } from '../../types';
 import { dagGraphParser } from '../parsers';
-import { DirectionType, GraphLayout, GraphLayoutBuilder, GraphTiers, TieredGraphLayoutBuilder } from './common';
+import { GraphLayout, GraphLayoutBuilder } from './common';
 
 class PlanarLayoutBuilder extends GraphLayoutBuilder<PlanarLayout> {
     _orientation: DirectionType = 'horizontal';
