+++ packages/ui-causal-graph-editor/src/shared/rendering/use-render-engine.tsx
@@ -14,12 +14,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { EdgeConstraint, EditorMode, SimulationEdge, SimulationGraph, ZoomThresholds } from '@types';
 import * as PIXI from 'pixi.js';
 import * as React from 'react';
 
 import { useTheme } from '@darajs/styled-components';
 import { NotificationPayload } from '@darajs/ui-notifications';
+
+import { EdgeConstraint, EditorMode, SimulationEdge, SimulationGraph, ZoomThresholds } from '@types';
 
 import { GraphLayout } from '../graph-layout';
 import { DragMode } from '../use-drag-mode';
