+++ packages/ui-causal-graph-editor/src/types.tsx
@@ -288,3 +288,21 @@ export interface ZoomThresholds {
 }
 
 export type ZoomState = Record<keyof ZoomThresholds, boolean>;
+
+export type DirectionType = 'horizontal' | 'vertical';
+
+export interface TiersConfig {
+    group: string;
+    order_nodes_by?: string;
+    rank?: string[];
+}
+
+export type GraphTiers = string[][] | TiersConfig;
+
+/**
+ * Defines necessary properties that need to be implemented by graph layouts to support tiered layouts
+ */
+export interface TieredGraphLayoutBuilder {
+    orientation: DirectionType;
+    tiers: GraphTiers;
+}
